change f32 to u64 in uncle rate target

+++ vm/types/src/on_chain_config/consensus.rs
@@ -7,7 +7,7 @@ use serde::{Deserialize, Serialize};
 /// Defines the version of Starcoin software.
 #[derive(Clone, Debug, Deserialize, PartialEq, Serialize)]
 pub struct Consensus {
-    pub uncle_rate_target: f32,
+    pub uncle_rate_target: u64,
     pub epoch_time_target: u64,
     pub reward_half_time_target: u64,
 }
